requirements.txt
Django>=5.0,<6
gTTS>=2.5
pydub>=0.25
Pillow>=10.0
nltk>=3.9
python-dotenv==1.0.1
grpcio==1.62.1
protobuf==5.27.2